output: remove support for multiple outputs
Multiple outputs has never really fit with Cage's window layouting model
and had several limitations. It also is difficult to unify with layer
shell surfaces, which expect to be run on specific outputs.

Until we find a way to improve the interaction between layer shell and
multiple outputs, remove it in favor of a proper layer shell
implementation.